+++ components/Dialogs/src/private/MDCAlertControllerView+Private.m
@@ -27,6 +27,7 @@ static const MDCFontTextStyle kMessageTextStyle = MDCFontTextStyleBody1;
 static const MDCFontTextStyle kButtonTextStyle = MDCFontTextStyleButton;
 
 static const CGFloat MDCDialogMaximumWidth = 560.0f;
+static const CGFloat MDCDialogMinimumWidth = 280.0f;
 
 static const CGFloat MDCDialogActionButtonMinimumHeight = 36.0f;
 static const CGFloat MDCDialogActionButtonMinimumWidth = 48.0f;
@@ -434,11 +435,11 @@ static const CGFloat MDCDialogMessageOpacity = 0.54f;
 }
 
 - (BOOL)hasTitleIconOrImage {
-  return self.titleIconImageView.image.size.height > 0.0f || self.titleIconView != nil;
+  return [self hasTitleIcon] || self.titleIconView != nil;
 }
 
-- (BOOL)hasTitleIconView {
-  return self.titleIconView != nil;
+- (BOOL)hasTitleIcon {
+  return self.titleIconImageView.image.size.height > 0.0f;
 }
 
 - (BOOL)hasTitle {
@@ -507,29 +508,47 @@ static const CGFloat MDCDialogMessageOpacity = 0.54f;
   return CGRectMake(self.contentInsets.left, top, messageSize.width, messageSize.height);
 }
 
-- (CGRect)titleIconFrameWithTitleSize:(CGSize)titleSize {
-  CGSize titleIconViewSize = [self titleIconViewSize];
-  CGRect titleFrame = [self titleFrameWithTitleSize:titleSize];
+/**
+ Calculate the frame of the titleIcon, given the title size and bounds size.
 
-  // match the titleIcon alignment to the title alignment
+ @param titleSize is the pre-calculated size of the title.
+ @param boundsSize is the total bounds without any internal margins or padding.
+*/
+- (CGRect)titleIconFrameWithTitleSize:(CGSize)titleSize boundsSize:(CGSize)boundsSize {
+  CGSize titleIconViewSize = [self titleIconViewSize];
   CGFloat leftInset = self.titleIconInsets.left;
   CGFloat topInset = self.titleIconInsets.top;
-  if (self.titleIconAlignment == NSTextAlignmentCenter) {
-    leftInset =
-        CGRectGetMinX(titleFrame) + (CGRectGetWidth(titleFrame) - titleIconViewSize.width) / 2.0f;
-  } else if (self.titleIconAlignment == NSTextAlignmentRight ||
-             (self.titleIconAlignment == NSTextAlignmentNatural &&
-              [self mdf_effectiveUserInterfaceLayoutDirection] ==
-                  UIUserInterfaceLayoutDirectionRightToLeft)) {
-    leftInset = CGRectGetMaxX(titleFrame) - titleIconViewSize.width;
+  CGFloat titleIconHeight = titleIconViewSize.height;
+  CGFloat titleIconWidth = titleIconViewSize.width;
+  BOOL isRightOrRTLNatural = (self.titleIconAlignment == NSTextAlignmentRight ||
+                              (self.titleIconAlignment == NSTextAlignmentNatural &&
+                               [self mdf_effectiveUserInterfaceLayoutDirection] ==
+                                   UIUserInterfaceLayoutDirectionRightToLeft));
+
+  if (self.titleIconAlignment == NSTextAlignmentJustified) {
+    // Justified images are sized to fit the alert's width (minus the insets).
+    titleIconWidth = boundsSize.width - leftInset - self.titleIconInsets.right;
+    if (titleIconViewSize.width > titleIconWidth && [self hasTitleIcon]) {
+      // If the width decreased, then decrease the height proportionally.
+      titleIconHeight *= titleIconWidth / titleIconViewSize.width;
+    }
+  } else {
+    CGRect titleFrame = [self titleFrameWithTitleSize:titleSize];
+    if (self.titleIconAlignment == NSTextAlignmentCenter) {
+      leftInset =
+          CGRectGetMinX(titleFrame) + (CGRectGetWidth(titleFrame) - titleIconViewSize.width) / 2.0f;
+    } else if (isRightOrRTLNatural) {
+      leftInset = CGRectGetMaxX(titleFrame) - titleIconViewSize.width;
+    }
   }
-  return CGRectMake(leftInset, topInset, titleIconViewSize.width, titleIconViewSize.height);
+
+  return CGRectMake(leftInset, topInset, titleIconWidth, titleIconHeight);
 }
 
 // @param boundsSize should not include any internal margins or padding
 - (CGSize)calculatePreferredContentSizeForBounds:(CGSize)boundsSize {
   // Even if we have more room, limit our maximum width
-  boundsSize.width = MIN(boundsSize.width, MDCDialogMaximumWidth);
+  boundsSize.width = MIN(MAX(boundsSize.width, MDCDialogMinimumWidth), MDCDialogMaximumWidth);
 
   // Title, Content & Actions
   CGSize titleViewSize = [self calculateTitleViewSizeThatFitsWidth:boundsSize.width];
@@ -538,7 +557,8 @@ static const CGFloat MDCDialogMessageOpacity = 0.54f;
 
   // Final Sizing
   CGSize totalSize;
-  totalSize.width = MAX(titleViewSize.width, MAX(contentSize.width, actionSize.width));
+  totalSize.width = MAX(MAX(titleViewSize.width, MAX(contentSize.width, actionSize.width)),
+                        MDCDialogMinimumWidth);
   totalSize.height = titleViewSize.height + contentSize.height + actionSize.height;
 
   return totalSize;
@@ -673,19 +693,21 @@ static const CGFloat MDCDialogMessageOpacity = 0.54f;
                               verticalFittingPriority:UILayoutPriorityFittingSizeLevel];
   accessoryViewSize.width = contentBoundsSize.width;
 
-  CGRect titleFrame = [self titleFrameWithTitleSize:titleSize];
-  CGRect messageFrame = [self messageFrameWithSize:messageSize];
-  CGRect accessoryViewFrame = CGRectMake(
-      self.contentInsets.left, CGRectGetMaxY(messageFrame) + [self accessoryVerticalInset],
-      accessoryViewSize.width, accessoryViewSize.height);
-
-  CGRect titleIconImageViewRect = [self titleIconFrameWithTitleSize:titleSize];
+  CGRect titleIconImageViewRect = [self titleIconFrameWithTitleSize:titleSize
+                                                         boundsSize:boundsSize];
   if (self.titleIconView != nil) {
     self.titleIconView.frame = titleIconImageViewRect;
   } else if (self.titleIconImageView != nil) {
     // Match the title icon alignment to the title alignment.
     self.titleIconImageView.frame = titleIconImageViewRect;
   }
+
+  // Calculate the title frame after the title icon size has been determined.
+  CGRect titleFrame = [self titleFrameWithTitleSize:titleSize];
+  CGRect messageFrame = [self messageFrameWithSize:messageSize];
+  CGRect accessoryViewFrame = CGRectMake(
+      self.contentInsets.left, CGRectGetMaxY(messageFrame) + [self accessoryVerticalInset],
+      accessoryViewSize.width, accessoryViewSize.height);
 
   self.titleLabel.frame = titleFrame;
   self.messageLabel.frame = messageFrame;