+++ components/FlexibleHeader/src/MDCFlexibleHeaderView.m
@@ -1301,28 +1301,14 @@ static inline MDCFlexibleHeaderShiftBehavior ShiftBehaviorForCurrentAppContext(
       self.transform = CGAffineTransformMakeTranslation(0, self.trackingScrollView.contentOffset.y);
     };
 
-    if (UIAccessibilityIsVoiceOverRunning()) {
-      // Clamp the offset to at least -self.maximumHeight. Accessibility may attempt to scroll to
-      // a lesser offset than this to pull the flexible header into the center of the scrollview on
-      // focusing.
-      CGPoint offset = self.trackingScrollView.contentOffset;
-      offset.y = MAX(offset.y, -self.minMaxHeight.maximumHeightWithTopSafeArea);
-      [self fhv_setContentOffset:offset forTrackingScrollView:self.trackingScrollView];
-      // Setting the transform on the same run loop as the accessibility scroll can cause additional
-      // incorrect scrolling as the scrollview attempts to resolve to a position that will place
-      // the header in the center of the scroll. Punting to the next loop prevents this.
-      dispatch_async(dispatch_get_main_queue(), ^{
-        updateTransform();
-        [self fhv_updateLayout];
-      });
-    } else {
+    CAAnimation *boundsAnimation = [self.trackingScrollView.layer animationForKey:@"bounds.origin"];
+
+    void (^updateTransformWithInFlightAnimation)(void) = ^{
       // Check if there is an in-flight bounds animation and piggy-back its duration in order to
       // avoid a jumping effect resulting from the transform otherwise updating instantly.
       // This can happen if one of the cells positioned above the app bar shrinks in height, which
       // causes the scroll view to animate its bounds origin in order to keep the scroll view's
       // content from moving.
-      CAAnimation *boundsAnimation =
-          [self.trackingScrollView.layer animationForKey:@"bounds.origin"];
       if (boundsAnimation) {
         [UIView animateWithDuration:boundsAnimation.duration
                          animations:^{
@@ -1331,6 +1317,36 @@ static inline MDCFlexibleHeaderShiftBehavior ShiftBehaviorForCurrentAppContext(
       } else {
         updateTransform();
       }
+    };
+
+    if (UIAccessibilityIsVoiceOverRunning()) {
+      // Clamp the offset to at least the max of -self.maximumHeight and the topContentInset.
+      // Accessibility may attempt to scroll to a lesser offset than this to pull the flexible
+      // header into the center of the scrollview on focusing.
+      CGPoint offset = self.trackingScrollView.contentOffset;
+      CGFloat scrollViewAdjustedContentInsetTop = self.trackingScrollView.contentInset.top;
+      if (@available(iOS 11.0, *)) {
+        scrollViewAdjustedContentInsetTop = self.trackingScrollView.adjustedContentInset.top;
+      }
+      offset.y = MAX(offset.y, -(MAX(self.minMaxHeight.maximumHeightWithTopSafeArea,
+                                     scrollViewAdjustedContentInsetTop)));
+      [self fhv_setContentOffset:offset forTrackingScrollView:self.trackingScrollView];
+      if (boundsAnimation) {
+        // The transform will piggy-back with the in-flight bounds
+        // animation in order to avoid a jumping effect.
+        updateTransformWithInFlightAnimation();
+      } else {
+        // Setting the transform on the same run loop as the accessibility scroll can cause
+        // additional incorrect scrolling as the scrollview attempts to resolve to a position that
+        // will place the header in the center of the scroll. Punting to the next loop prevents
+        // this.
+        dispatch_async(dispatch_get_main_queue(), ^{
+          updateTransform();
+          [self fhv_updateLayout];
+        });
+      }
+    } else {
+      updateTransformWithInFlightAnimation();
     }
   }