+++ components/Tabs/src/private/MDCItemBar.m
@@ -60,6 +60,11 @@ static void *kItemPropertyContext = &kItemPropertyContext;
 
 #pragma mark -
 
+#ifdef __IPHONE_13_4
+@interface MDCItemBar (PointerInteraction) <UIPointerInteractionDelegate>
+@end
+#endif
+
 @interface MDCItemBar () <UICollectionViewDataSource, UICollectionViewDelegateFlowLayout>
 // Current style properties.
 @property(nonatomic, strong, nullable) MDCItemBarStyle *style;
@@ -569,6 +574,18 @@ static void *kItemPropertyContext = &kItemPropertyContext;
   return [NSIndexPath indexPathForItem:index inSection:0];
 }
 
+- (void)invalidateItemCellPointerInteractions {
+#ifdef __IPHONE_13_4
+  if (@available(iOS 13.4, *)) {
+    for (MDCItemBarCell *cell in self.collectionView.visibleCells) {
+      for (UIPointerInteraction *interaction in cell.interactions) {
+        [interaction invalidate];
+      }
+    }
+  }
+#endif
+}
+
 - (void)reload {
   [_collectionView reloadData];
   [self updateAlignmentAnimated:NO];
@@ -593,6 +610,10 @@ static void *kItemPropertyContext = &kItemPropertyContext;
     [self->_selectionIndicator layoutIfNeeded];
   };
 
+  void (^completionBlock)(void) = ^{
+    [self invalidateItemCellPointerInteractions];
+  };
+
   if (animate) {
     CAMediaTimingFunction *easeInOutFunction =
         [CAMediaTimingFunction mdc_functionWithType:MDCAnimationTimingFunctionEaseInOut];
@@ -600,6 +621,7 @@ static void *kItemPropertyContext = &kItemPropertyContext;
     [CATransaction begin];
     [CATransaction setAnimationDuration:kDefaultAnimationDuration];
     [CATransaction setAnimationTimingFunction:easeInOutFunction];
+    [CATransaction setCompletionBlock:completionBlock];
     [UIView animateWithDuration:kDefaultAnimationDuration
                           delay:0
                         options:UIViewAnimationOptionBeginFromCurrentState
@@ -609,6 +631,7 @@ static void *kItemPropertyContext = &kItemPropertyContext;
 
   } else {
     animationBlock();
+    completionBlock();
   }
 }
 
@@ -825,6 +848,17 @@ static void *kItemPropertyContext = &kItemPropertyContext;
 - (void)configureCell:(MDCItemBarCell *)cell {
   // Configure content style
   [cell applyStyle:_style];
+
+#ifdef __IPHONE_13_4
+  if (@available(iOS 13.4, *)) {
+    // Add a pointer interaction if necessary
+    if (cell.interactions.count == 0) {
+      UIPointerInteraction *pointerInteraction =
+          [[UIPointerInteraction alloc] initWithDelegate:self];
+      [cell addInteraction:pointerInteraction];
+    }
+  }
+#endif
 }
 
 - (void)configureVisibleCells {
@@ -851,6 +885,21 @@ static void *kItemPropertyContext = &kItemPropertyContext;
 - (void)updateSelectionIndicatorVisibility {
   _selectionIndicator.hidden = !_style.shouldDisplaySelectionIndicator;
 }
+
+#pragma mark - UIPointerInteractionDelegate
+
+#ifdef __IPHONE_13_4
+- (UIPointerStyle *)pointerInteraction:(UIPointerInteraction *)interaction
+                        styleForRegion:(UIPointerRegion *)region API_AVAILABLE(ios(13.4)) {
+  UIPointerStyle *pointerStyle = nil;
+  if (interaction.view) {
+    UITargetedPreview *targetedPreview = [[UITargetedPreview alloc] initWithView:interaction.view];
+    UIPointerEffect *highlightEffect = [UIPointerHighlightEffect effectWithPreview:targetedPreview];
+    pointerStyle = [UIPointerStyle styleWithEffect:highlightEffect shape:nil];
+  }
+  return pointerStyle;
+}
+#endif
 
 @end